Refactor out appbar and sidemenu from App.js into components/
